requirements.txt
torch==1.13.1
torchvision==0.14.1
accelerate==0.16.0
clip==0.1.0
datasets==2.9.0
diffusers==0.14.0
fsspec==2022.11.0
fairscale==0.4.4
ftfy==6.1.1
huggingface-hub==0.14.0
iopath==0.1.10
jsonlines==3.1.0
levenshtein==0.20.9
nltk==3.8.1
omegaconf==2.3.0
opencv-python==4.7.0.68
pycocoevalcap==1.2
pycocotools==2.0.6
python-levenshtein==0.20.9
ruamel-yaml==0.17.21
scikit-image==0.19.3
scikit-learn==1.0
scipy==1.10.0
sentence-transformers==2.2.2
tensorboard==2.11.2
timm==0.4.12
tokenizers==0.13.2
torch-fidelity==0.3.0
torch-utils==0.1.2
tqdm==4.63.2
